chore: compress whitespaces in Driver.textIs
It is experimental, making it similar to isInnerHTML method, do not change other methods like textIsAll for now. The reason for the change is Safari returns original DOM for a text unlike Chrome and Firefox, so jamming whitespaces into a single space seems working for all browsers. You cannot test whitespaced text though. If you ever need it, then { trim } options should be probably added to textIs() method to control the behavior.